+++ Tasks/CodeQuality/PesterIntegration.ps1
@@ -37,10 +37,11 @@ Add-BuildTask -Name PesterIntegration -Jobs SetVariables, {
         return
     }
 
-    $config = New-PesterConfiguration
-    $config.Run.Path = "$BuildRoot\Tests\Integration"
-    $config.Run.PassThru = $true
-    $result = Invoke-Pester -Configuration $config
+    $pesterSplat = @{
+        Path           = "$BuildRoot\Tests\Integration"
+        ModuleManifest = Join-Path $BuildRoot "$script:moduleName.psd1"
+    }
+    $result = Invoke-PlumberPester @pesterSplat
 
     $failures = $result | Where-Object { $_.Result -eq 'Failed' }
     if ($failures) { Write-Error "Pester failed with $($failures.count) error(s)" }